requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
